39 Python3 Web框架/3 FastAPI/project/config/config-local/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pydantic-settings
python-dotenv
pyyaml
